+++ bdns/problem_test.go
@@ -1,4 +1,9 @@
-package dns
+// Copyright 2015 ISRG.  All rights reserved
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+package bdns
 
 import (
 	"errors"
